Amend Stock Order Limit
- Add number of days of sales input
- Add number of days of sales input
- Amend percentage decimal

+++ StockManagementSystem.Services/OrderLimits/OrderLimitService.cs
@@ -28,7 +28,7 @@ namespace StockManagementSystem.Services.OrderLimits
         
         public Task<IPagedList<OrderLimit>> GetOrderLimitsAsync(
             int[] storeIds = null,
-            float? percentage = 0,
+            int? percentage = 0,
             int pageIndex = 0,
             int pageSize = int.MaxValue,
             bool getOnlyTotalCount = false)
